+++ src/domain/auth-rebac/services/src/dependencies.rs
@@ -19,6 +19,7 @@ pub fn register_dependencies(b: &mut CatalogBuilder, needs_indexing: bool) {
     }
     b.add::<RebacDatasetLifecycleMessageConsumer>();
     b.add::<RebacServiceImpl>();
+    b.add::<RebacDatasetRegistryFacadeImpl>();
     b.add_value(DefaultAccountProperties { is_admin: false });
     b.add_value(DefaultDatasetProperties {
         allows_anonymous_read: false,