+++ src/domain/auth-rebac/services/src/lib.rs
@@ -14,6 +14,7 @@ mod dependencies;
 mod jobs;
 mod messages;
 mod rebac_dataset_lifecycle_message_consumer;
+mod rebac_dataset_registry_facade_impl;
 mod rebac_indexer;
 mod rebac_service_impl;
 
@@ -21,5 +22,6 @@ pub use dependencies::*;
 pub use jobs::*;
 pub use messages::*;
 pub use rebac_dataset_lifecycle_message_consumer::*;
+pub use rebac_dataset_registry_facade_impl::*;
 pub use rebac_indexer::*;
 pub use rebac_service_impl::*;