+++ src/domain/datasets/services/src/utils/mod.rs
@@ -7,7 +7,6 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-pub(crate) mod access_dataset_helper;
 pub mod create_dataset_use_case_helper;
 
 pub use create_dataset_use_case_helper::*;