+++ src/infra/core/src/use_cases/set_watermark_use_case_impl.rs
@@ -11,8 +11,9 @@ use std::sync::Arc;
 
 use chrono::{DateTime, Utc};
 use dill::{component, interface};
-use kamu_core::auth::{DatasetAction, DatasetActionAuthorizer};
-use kamu_core::*;
+use internal_error::ErrorIntoInternal;
+use kamu_auth_rebac::RebacDatasetRegistryFacade;
+use kamu_core::{auth, *};
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -21,22 +22,19 @@ use kamu_core::*;
 pub struct SetWatermarkUseCaseImpl {
     set_watermark_planner: Arc<dyn SetWatermarkPlanner>,
     set_watermark_executor: Arc<dyn SetWatermarkExecutor>,
-    dataset_registry: Arc<dyn DatasetRegistry>,
-    dataset_action_authorizer: Arc<dyn DatasetActionAuthorizer>,
+    rebac_dataset_registry_facade: Arc<dyn RebacDatasetRegistryFacade>,
 }
 
 impl SetWatermarkUseCaseImpl {
     pub fn new(
         set_watermark_planner: Arc<dyn SetWatermarkPlanner>,
         set_watermark_executor: Arc<dyn SetWatermarkExecutor>,
-        dataset_registry: Arc<dyn DatasetRegistry>,
-        dataset_action_authorizer: Arc<dyn DatasetActionAuthorizer>,
+        rebac_dataset_registry_facade: Arc<dyn RebacDatasetRegistryFacade>,
     ) -> Self {
         Self {
             set_watermark_planner,
             set_watermark_executor,
-            dataset_registry,
-            dataset_action_authorizer,
+            rebac_dataset_registry_facade,
         }
     }
 }
@@ -55,16 +53,18 @@ impl SetWatermarkUseCase for SetWatermarkUseCaseImpl {
         dataset_handle: &odf::DatasetHandle,
         new_watermark: DateTime<Utc>,
     ) -> Result<SetWatermarkResult, SetWatermarkError> {
-        // Permission check
-        self.dataset_action_authorizer
-            .check_action_allowed(&dataset_handle.id, DatasetAction::Write)
-            .await?;
-
         // Resolve dataset
         let target = self
-            .dataset_registry
-            .get_dataset_by_handle(dataset_handle)
-            .await;
+            .rebac_dataset_registry_facade
+            .resolve_dataset_by_handle(dataset_handle, auth::DatasetAction::Write)
+            .await
+            .map_err(|e| {
+                use kamu_auth_rebac::RebacDatasetIdUnresolvedError as E;
+                match e {
+                    E::Access(e) => SetWatermarkError::Access(e),
+                    e @ E::Internal(_) => SetWatermarkError::Internal(e.int_err()),
+                }
+            })?;
 
         // Make a plan
         let plan = self